requirements.txt
certifi
cryptography
chardet
click
future
idna
itsdangerous
Jinja2
MarkupSafe
oauthlib
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
urllib3
flask==1.1.1
webexteamssdk==1.6
PyJWT==1.7.1.
python-dotenv==0.19.0
